Add files via upload
This project for predicting the if the space x falcon 9 rocket will land successfully or not 
first of all i should talk in a brief for space x falcon 9 rocket launch , Space X advertises Falcon 9 rocket launches on its website with a cost of 62 million dollars; other providers cost upward of 165 million dollars each, much of the savings is because Space X can reuse the first stage. Therefore if i can determine if the first stage will land, i can determine the cost of a launch. This information can be used if an alternate company wants to bid against space X for a rocket launch.
So i first should extract and get the proper data for my project and there are two ways webscraping or by APIs ; the first way the webscraped wikipedia page for falcon 9 past launches and the second way is using SapceX API to collect the past launches then save this data to csv file to use in the next steps .
Second the step of wrangling the data and make it proper for later analysis and modeling by removing any missing values or transforming the data to proper format and i made column called class by analyzing column Output to clarify if it successfully land or not .
third i made some expletory data analysis to show some insights in our data and answer some non answered question by using SQL queries 
fourth i started implementing  a classification model (SVM , KNN , Decision Tree and logistic regression) to predict the falcon 9 first stage landing will be successful of not